+++ R/as_flextable.R
@@ -302,9 +302,6 @@ as_flextable.lm <- function(x, ...){
 #'   ft_1 <- as_flextable(chisq.test(M))
 #'   ft_1
 #' }
-#' @section Illustrations:
-#'
-#' \if{html}{\figure{fig_as_flextable.htest_1.png}{options: width="500"}}
 #' @family as_flextable methods
 as_flextable.htest <- function (x, ...) {
   ret <- x[c("estimate", "statistic", "p.value", "parameter")]
@@ -369,9 +366,6 @@ as_flextable.htest <- function (x, ...) {
 #' ft_1 <- continuous_summary(iris, names(iris)[1:4], by = "Species",
 #'   hide_grouplabel = FALSE)
 #' ft_1
-#' @section Illustrations:
-#'
-#' \if{html}{\figure{fig_continuous_summary_1.png}{options: width="500"}}
 continuous_summary <- function(dat, columns = NULL,
                                by = character(0),
                                hide_grouplabel = TRUE,
@@ -444,6 +438,7 @@ continuous_summary <- function(dat, columns = NULL,
 #'   ft <- as_flextable(m1)
 #'   ft
 #' }
+#' @family as_flextable methods
 as_flextable.merMod <- function(x, ...){
 
   if( !requireNamespace("broom.mixed", quietly = TRUE) ){
@@ -551,6 +546,7 @@ as_flextable.glmmadmb <- as_flextable.merMod
 #'   ft
 #' }
 #' @importFrom rlang sym
+#' @family as_flextable methods
 as_flextable.kmeans <- function(x, digits = 4, ...) {
   if (!requireNamespace("broom", quietly = TRUE)) {
     stop(sprintf(
@@ -671,6 +667,7 @@ as_flextable.kmeans <- function(x, digits = 4, ...) {
 #'   ft <- as_flextable(cl)
 #'   ft
 #' }
+#' @family as_flextable methods
 as_flextable.pam <- function(x, digits = 4, ...){
   if( !requireNamespace("broom", quietly = TRUE) ){
     if (!requireNamespace("broom", quietly = TRUE)) {
@@ -762,6 +759,189 @@ as_flextable.pam <- function(x, digits = 4, ...){
 }
 
 #' @export
+#' @title Printing data.frame summary as a flextable
+#' @description It displays the first rows and shows the column types.
+#' If there is only one row, a simplified vertical table is produced.
+#' @param x a data.frame
+#' @param max_row The number of rows to print. Default to 10.
+#' @param split_colnames Should the column names be split
+#' (with non alpha-numeric characters). Default to FALSE.
+#' @param short_strings Should the character column be shorten.
+#' Default to FALSE.
+#' @param short_size Maximum length of character column if
+#' `short_strings` is TRUE. Default to 35.
+#' @param short_suffix Suffix to add when character values are shorten.
+#' Default to "...".
+#' @param do_autofit Use [autofit()] before rendering the table.
+#' Default to TRUE.
+#' @param show_coltype Show column types.
+#' Default to TRUE.
+#' @param color_coltype Color to use for column types.
+#' Default to "#999999".
+#' @param ... unused arguments
+#' @examples
+#' as_flextable(mtcars)
+#' @family as_flextable methods
+as_flextable.data.frame <- function(x,
+                                    max_row = 10,
+                                    split_colnames = FALSE,
+                                    short_strings = FALSE,
+                                    short_size = 35,
+                                    short_suffix = "...",
+                                    do_autofit = TRUE,
+                                    show_coltype = TRUE,
+                                    color_coltype = "#999999",
+                                    ...) {
+  if (nrow(x) == 1) {
+    singlerow_df_printer(
+      dat = x,
+      max_row = max_row,
+      short_strings = short_strings,
+      short_size = short_size,
+      short_suffix = short_suffix,
+      do_autofit = do_autofit,
+      show_coltype = show_coltype,
+      color_coltype = color_coltype
+    )
+  } else {
+    multirow_df_printer(
+      dat = x,
+      max_row = max_row,
+      split_colnames = split_colnames,
+      short_strings = short_strings,
+      short_size = short_size,
+      short_suffix = short_suffix,
+      do_autofit = do_autofit,
+      show_coltype = show_coltype,
+      color_coltype = color_coltype
+    )
+  }
+}
+
+look_like_int <- function(x) {
+  (is.numeric(x) && isTRUE(all.equal(x, as.integer(x)))) || is.integer(x)
+}
+
+multirow_df_printer <- function(dat,
+                                max_row = 10,
+                                split_colnames = FALSE,
+                                short_strings = FALSE,
+                                short_size = 35,
+                                short_suffix = "...",
+                                do_autofit = TRUE,
+                                show_coltype = TRUE,
+                                color_coltype = "#999999") {
+  x <- as.data.frame(dat)
+  nro <- nrow(x)
+
+  z <- get_flextable_defaults()
+
+  x <- head(x, n = max_row)
+  coltypes <- as.character(sapply(dat, function(x) head(class(x), 1)))
+
+  lli <- sapply(x, look_like_int)
+  x[lli] <- lapply(x[lli], as.integer)
+
+  if (!is.null(short_strings) && short_strings) {
+    wic <- sapply(x, is.character)
+    x[wic] <- lapply(x[wic], function(x) {
+      paste0(substring(text = x, first = 1, last = short_size), short_suffix)
+    })
+  }
+
+  colkeys <- colnames(x)
+
+  ft <- flextable(x, col_keys = colkeys)
+
+  if (split_colnames) {
+    labs <- strsplit(colkeys, split = "[^[:alnum:]]+")
+    names(labs) <- colkeys
+    labs <- lapply(labs, paste, collapse = "\n")
+    ft <- set_header_labels(ft, values = labs)
+  }
+
+  if (show_coltype) {
+    ft <- add_header_row(ft, top = FALSE, values = coltypes)
+  }
+
+  ft <- colformat_double(ft)
+  ft <- colformat_int(ft)
+
+  if (nro > max_row) {
+    ft <- add_footer_lines(ft, values = sprintf("n: %.0f", nro))
+  }
+  ft <- set_table_properties(ft, layout = z$table.layout)
+  if ("fixed" %in% z$table.layout && do_autofit) {
+    ft <- autofit(ft)
+  }
+
+  ft <- do.call(z$theme_fun, list(ft))
+
+  if (show_coltype) {
+    ft <- color(ft, i = nrow_part(ft, "header"), part = "header", color = color_coltype)
+  }
+  ft <- align(ft, align = "left", part = "footer")
+  ft
+}
+
+singlerow_df_printer <- function(dat,
+                                 max_row = 10,
+                                 split_colnames = FALSE,
+                                 short_strings = FALSE,
+                                 short_size = 35,
+                                 short_suffix = "...",
+                                 do_autofit = TRUE,
+                                 show_coltype = TRUE,
+                                 color_coltype = "#999999") {
+  coltypes <- as.character(sapply(dat, function(x) head(class(x), 1)))
+
+  lli <- sapply(dat, look_like_int)
+  dat[lli] <- lapply(dat[lli], as.integer)
+
+  x <- data.frame(
+    "Col." = colnames(dat),
+    "Type" = coltypes,
+    "Val." = vapply(dat, format_fun.default, FUN.VALUE = NA_character_)
+  )
+
+  z <- get_flextable_defaults()
+
+  x <- head(x, n = max_row)
+
+  if (!is.null(short_strings) && short_strings) {
+    wic <- sapply(x, is.character)
+    x[wic] <- lapply(x[wic], function(x) {
+      paste0(substring(text = x, first = 1, last = short_size), short_suffix)
+    })
+  }
+
+  colkeys <- c("Col.", "Val.")
+
+  ft <- flextable(x, col_keys = colkeys)
+  ft <- delete_part(ft, part = "header")
+  ft <- set_table_properties(ft, layout = z$table.layout)
+  if ("fixed" %in% z$table.layout && do_autofit) {
+    ft <- autofit(ft)
+  }
+  ft <- do.call(z$theme_fun, list(ft))
+  ft <- align(ft, align = "center", part = "all")
+  ft <- align(ft, j = 1, align = "right", part = "all")
+  ft <- valign(x = ft, valign = "top", part = "body")
+  if (show_coltype) {
+    ft <- append_chunks(
+      x = ft, j = "Col.",
+      as_chunk("\n"),
+      colorize(
+        x = as_chunk(Type, props = fp_text_default(font.size = z$font.size * 2 / 3)),
+        color = color_coltype
+      )
+    )
+  }
+  ft
+}
+
+
+#' @export
 #' @title set model automatic printing as a flextable
 #' @description Define [as_flextable()] as
 #' print method in an R Markdown document for models
@@ -799,7 +979,7 @@ use_model_printer <- function() {
   invisible()
 }
 
-utils::globalVariables(c("p.value", "value"))
+utils::globalVariables(c("p.value", "value", "Type"))