+++ R/themes.R
@@ -168,7 +168,8 @@ theme_vader <- function(x, ...) {
   if (h_nrow > 0) {
     x <- hline_bottom(x, border = big_border, part = "header")
   }
-  if (b_nrow > 0) {
+
+  if (b_nrow > 0 || h_nrow < 1) {
     x <- hline_top(x, border = big_border, part = "body")
   }
   if (f_nrow > 0) {
@@ -360,7 +361,12 @@ theme_booktabs <- function(x, bold_header = FALSE, ...) {
     x <- hline(x, border = std_border, part = "header")
     x <- hline_bottom(x, border = big_border, part = "header")
     x <- bold(x, bold = bold_header, part = "header")
+  } else if (b_nrow > 0) {
+    x <- hline_top(x, border = big_border, part = "body")
+  } else if (f_nrow > 0) {
+    x <- hline_top(x, border = big_border, part = "body")
   }
+
   if (f_nrow > 0) {
     x <- hline_bottom(x, border = big_border, part = "footer")
   }
@@ -385,9 +391,6 @@ theme_booktabs <- function(x, bold_header = FALSE, ...) {
 #' ft <- flextable(head(mtcars*22.22))
 #' ft <- theme_apa(ft)
 #' ft
-#' @section Illustrations:
-#'
-#' \if{html}{\figure{fig_theme_apa_1.png}{options: width="500"}}
 theme_apa <- function(x, ...) {
   if (!inherits(x, "flextable")) {
     stop(sprintf("Function `%s` supports only flextable objects.", "theme_apa()"))